requirements.txt
fastapi[standard]>=0.135.1
sqlalchemy
pandas
langchain>=1.2.12
langchain-ollama>=1.0.1
markdown-pdf>=1.13.1
matplotlib>=3.10.8
numpy>=2.4.3
seaborn>=0.13.2
httpx
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
alembic
pydantic-settings
sqlalchemy-utils
deep-translator
fastapi-cache2
